+++ ddtrace/contrib/internal/langchain/utils.py
@@ -1,7 +1,7 @@
 import inspect
 import sys
 
-from ddtrace.settings.asm import config as asm_config
+from ddtrace.internal import core
 
 
 class BaseTracedLangChainStreamResponse:
@@ -99,7 +99,7 @@ def shared_stream(
     try:
         resp = func(*args, **kwargs)
         cls = TracedLangchainAsyncStreamResponse if inspect.isasyncgen(resp) else TracedLangchainStreamResponse
-        chunk_callback = _iast_create_chunk_callback(args, kwargs)
+        chunk_callback = _get_chunk_callback(interface_type, args, kwargs)
         return cls(resp, integration, span, on_span_finished, chunk_callback)
     except Exception:
         # error with the method call itself
@@ -134,93 +134,26 @@ def tag_general_message_input(span, inputs, integration, langchain_core):
             span.set_tag_str("langchain.request.messages.%d.content" % (input_idx), integration.trunc(str(inp)))
 
 
-def _iast_chunk_no_op(chunk):
-    pass
+def _get_chunk_callback(interface_type, args, kwargs):
+    results = core.dispatch_with_results("langchain.stream.chunk.callback", (interface_type, args, kwargs))
+    callbacks = []
+    for result in results.values():
+        if result and result.value:
+            callbacks.append(result.value)
+    return _build_chunk_callback(callbacks)
 
 
-def _iast_create_taint_chunk_callback(source):
-    def _iast_chunk_taint(chunk):
-        _iast_taint_chunk(source, chunk)
+def _build_chunk_callback(callbacks):
+    if not callbacks:
+        return _no_op_callback
+
+    def _chunk_callback(chunk):
+        for callback in callbacks:
+            callback(chunk)
         return chunk
 
-    return _iast_chunk_taint
+    return _chunk_callback
 
 
-def _iast_create_chunk_callback(args, kwargs):
-    from ddtrace.internal.utils import get_argument_value
-
-    chat_messages = get_argument_value(args, kwargs, 0, "input", optional=True)
-    if not chat_messages:
-        return _iast_chunk_no_op
-    source = _iast_get_tainted_source_from_chat_prompt_value(chat_messages)
-    if not source:
-        return _iast_chunk_no_op
-    return _iast_create_taint_chunk_callback(source)
-
-
-def _iast_get_tainted_source_from_chat_prompt_value(chat_prompt_value):
-    if not asm_config._iast_enabled:
-        return None
-    if not hasattr(chat_prompt_value, "messages"):
-        return None
-    messages = chat_prompt_value.messages
-    if not isinstance(messages, (tuple, list)):
-        return None
-
-    from ddtrace.appsec._iast._taint_tracking._taint_objects import get_tainted_ranges
-
-    for message in messages:
-        if not hasattr(message, "content"):
-            continue
-        content = message.content
-        if not isinstance(content, str):
-            continue
-        tainted_ranges = get_tainted_ranges(content)
-        if tainted_ranges:
-            return tainted_ranges[0].source
-    return None
-
-
-def _iast_taint_chunk(source, chunk):
-    """
-    Taints a chunk (type BaseMessageChunk, typically an AIMessageChunk) given a source.
-    """
-    # Relevant attributes to taint are:
-    #  content: Union[str, list[Union[str, dict]]]
-    #  additional_kwargs: dict
-    if not asm_config._iast_enabled:
-        return
-    if not source:
-        return
-    message = chunk
-    if not hasattr(message, "content"):
-        return
-    content = message.content
-    if isinstance(content, str):
-        setattr(message, "content", _iast_taint_if_str(source, content))
-    elif isinstance(content, list):
-        setattr(message, "content", [_iast_taint_if_str(source, c) for c in content])
-    elif isinstance(content, dict):
-        setattr(message, "content", {k: _iast_taint_if_str(source, v) for k, v in message.items()})
-    if hasattr(message, "additional_kwargs"):
-        additional_kwargs = message.additional_kwargs
-        if isinstance(additional_kwargs, dict) and "function_call" in additional_kwargs:
-            # OpenAI-style tool call, arguments are passed serialized in JSON.
-            function_call = additional_kwargs["function_call"]
-            if isinstance(function_call, dict) and "arguments" in function_call:
-                arguments = function_call["arguments"]
-                if isinstance(arguments, str):
-                    function_call["arguments"] = _iast_taint_if_str(source, arguments)
-
-
-def _iast_taint_if_str(source, obj):
-    if not isinstance(obj, str):
-        return obj
-    from ddtrace.appsec._iast._taint_tracking._taint_objects import taint_pyobject
-
-    return taint_pyobject(
-        pyobject=obj,
-        source_name=source.name,
-        source_value=source.value,
-        source_origin=source.origin,
-    )
+def _no_op_callback(chunk):
+    pass
